Modify SimMotor to use the same FlyMotorInfo as the EPICS motor, move FlyMotorInfo to core
